Use TS, start updater system.
The updater system will run through a series of "steps", each of which prompt the user, and then may apply changes depending on responses.

This is a first iteration; future versions will probably be cleaner.
